Cleaning up, added Footer
