+++ src/components/Vis1Explanation.jsx
