+++ src/pages/App.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react'
 import { Container } from 'semantic-ui-react'
 import HeaderSection from './HeaderSection'
+import FooterSection from './FooterSection'
 
 class App extends Component {
   constructor (props) {
@@ -16,7 +17,7 @@ class App extends Component {
       <Container fluid >
         <HeaderSection />
         {this.props.children}
-
+        <FooterSection />
       </Container>
     )
   }
