+++ src/pages/StateMap.jsx
@@ -77,9 +77,9 @@ export default class PAMap extends Component {
                   <path
                     key={ `path-${ i }` }
                     d={ path(d) }
-                    className="county"
-                    fill={ `rgb(40,40,40)` }
-                    stroke="#666"
+                    className='county'
+                    fill={ 'rgb(40,40,40)' }
+                    stroke='#666'
                     strokeWidth={ 1 }
                   />
                   <text
