+++ src/routes.js
@@ -3,7 +3,6 @@ import { Router, Route, IndexRoute } from 'react-router'
 
 import App from './pages/App.jsx'
 import Landing from './pages/Landing.jsx'
-import Visualization3 from './pages/Visualization3.jsx'
 import County from './pages/County.jsx'
 import School from './pages/School.jsx'
 
@@ -11,7 +10,6 @@ export default (
   <Router>
     <Route path='/' component={App}>
       <IndexRoute component={Landing} />
-      <Route path='visualization' component={Visualization3} />
       <Route path='county/:county' component={County} />
       <Route path='school/:school' component={School} />
       <Route path='*' component={Landing} />
