+++ src/styles/index.sass
@@ -227,3 +227,16 @@ svg
      padding-bottom: 0px !important
    .district-section
      padding-bottom: 0px !important
+
+.footer-container
+  background-color: rgba(40, 40, 40, .9)
+  color: #fff
+  font-weight: normal !important
+  font-family: 'Avenir Next' !important
+  padding: 5px
+  margin-top: 20px
+
+  .footer-content
+
+    a
+      color: #fff !important
